backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
django-cors-headers
django-ratelimit==4.1.0
django-redis==5.4.0
python-dotenv==1.0.1

